requirements_web.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.3
openai==1.36.1
python-dotenv==1.0.1
celery==5.3.4
redis==5.0.1
